# apps/admin-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/semester": {
      "post": {
        "operationId": "SemesterController_createSemester",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSemesterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "새로운 학기를 생성합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterEntity"
                }
              }
            }
          },
          "400": {
            "description": "학기 생성에 실패했습니다."
          }
        },
        "summary": "새로운 학기 생성",
        "tags": [
          "Semester"
        ]
      },
      "get": {
        "operationId": "SemesterController_getAllSemesters",
        "parameters": [],
        "responses": {
          "200": {
            "description": "모든 학기를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SemesterEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "학기를 조회하는 데 실패했습니다."
          }
        },
        "summary": "모든 학기 목록 조회",
        "tags": [
          "Semester"
        ]
      }
    },
    "/semester/{id}": {
      "get": {
        "operationId": "SemesterController_getSemesterById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 ID를 가진 학기를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterEntity"
                }
              }
            }
          },
          "400": {
            "description": "학기를 조회하는 데 실패했습니다."
          },
          "404": {
            "description": "학기를 찾을 수 없습니다."
          }
        },
        "summary": "특정 학기 조회",
        "tags": [
          "Semester"
        ]
      },
      "patch": {
        "operationId": "SemesterController_updateSemester",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSemesterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "특정 ID를 가진 학기를 업데이트합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterEntity"
                }
              }
            }
          },
          "400": {
            "description": "학기 업데이트에 실패했습니다."
          },
          "404": {
            "description": "업데이트하려는 학기를 찾을 수 없습니다."
          }
        },
        "summary": "특정 학기 수정",
        "tags": [
          "Semester"
        ]
      },
      "delete": {
        "operationId": "SemesterController_deleteSemester",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 ID를 가진 학기를 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemesterEntity"
                }
              }
            }
          },
          "400": {
            "description": "학기 삭제에 실패했습니다."
          },
          "404": {
            "description": "삭제하려는 학기를 찾을 수 없습니다."
          }
        },
        "summary": "특정 학기 삭제",
        "tags": [
          "Semester"
        ]
      }
    },
    "/lecture": {
      "post": {
        "operationId": "LectureController_createLecture",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLectureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "새로운 강의를 생성합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureEntity"
                }
              }
            }
          },
          "400": {
            "description": "강의 생성에 실패했습니다."
          }
        },
        "summary": "새로운 강의 생성",
        "tags": [
          "Lecture"
        ]
      },
      "get": {
        "operationId": "LectureController_getLectures",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "query",
            "description": "조회할 강의가 속한 학기의 연도입니다.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": true,
            "in": "query",
            "description": "조회할 강의가 속한 학기의 학기(예: Summer, Winter)입니다.",
            "schema": {
              "enum": [
                "Spring",
                "Summer",
                "Fall",
                "Winter"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 학기와 학기에 속한 강의를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LectureEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "강의를 조회하는 데 실패했습니다."
          }
        },
        "summary": "특정 학기의 강의 목록 조회",
        "tags": [
          "Lecture"
        ]
      }
    },
    "/lecture/{id}": {
      "patch": {
        "operationId": "LectureController_updateLecture",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLectureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "특정 ID를 가진 강의를 업데이트합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureEntity"
                }
              }
            }
          },
          "400": {
            "description": "강의 업데이트에 실패했습니다."
          },
          "404": {
            "description": "업데이트하려는 강의를 찾을 수 없습니다."
          }
        },
        "summary": "특정 강의 수정",
        "tags": [
          "Lecture"
        ]
      },
      "delete": {
        "operationId": "LectureController_deleteLecture",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 ID를 가진 강의를 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureEntity"
                }
              }
            }
          },
          "400": {
            "description": "강의 삭제에 실패했습니다."
          },
          "404": {
            "description": "삭제하려는 강의를 찾을 수 없습니다."
          }
        },
        "summary": "특정 강의 삭제",
        "tags": [
          "Lecture"
        ]
      }
    },
...
    "/student-lecture/semester/{year}/{season}": {
      "get": {
        "operationId": "StudentLectureController_getStudentsWithLectureLevelsBySemester",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "season",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "해당 학기에 수강 중인 학생 목록을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentEntity"
                  }
                }
              }
            }
          },
          "400": {
            "description": "학생울 조회하는 데 실패했습니다."
          }
        },
        "summary": "특정 학기의 수강생 목록 조회",
        "tags": [
          "StudentLecture"
        ]
      }
    },
    "/student-lecture": {
      "post": {
        "operationId": "StudentLectureController_createStudentWithLectureLog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentLectureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "새로운 학생과 강의 로그를 생성합니다."
          },
          "400": {
            "description": "학생 생성에 실패했습니다."
          }
        },
        "summary": "새로운 학생과 강의 로그 생성",
        "tags": [
          "StudentLecture"
        ]
      }
    },
    "/student-lecture/multiple": {
      "post": {
        "operationId": "StudentLectureController_createMultipleStudents",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "새로운 여러 학생과 강의 로그를 생성합니다."
          },
          "400": {
            "description": "학생 생성에 실패했습니다."
          }
        },
        "summary": "새로운 여러 학생과 강의 로그 생성",
        "tags": [
          "StudentLecture"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sinchon Admin Backend",
    "description": "신촌연합 관리자 페이지 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateSemesterDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "minimum": 1,
            "description": "해당 학기가 속한 연도 (예: 2024)",
            "example": 2024
          },
          "season": {
            "type": "string",
            "description": "해당 학기가 속한 계절 (Spring, Summer, Fall, Winter 중 하나)",
            "enum": [
              "Spring",
              "Summer",
              "Fall",
              "Winter"
            ],
            "example": "Summer"
          }
        },
        "required": [
          "year",
          "season"
        ]
      },
      "SemesterEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "고유 ID 값"
          },
          "year": {
            "type": "number",
            "description": "해당 학기가 속한 연도. ex) 2024"
          },
          "season": {
            "type": "string",
            "enum": [
              "Spring",
              "Summer",
              "Fall",
              "Winter"
            ],
            "description": "해당 학기가 속한 계절. ex) Summer"
          }
        },
        "required": [
          "id",
          "year",
          "season"
        ]
      },
      "UpdateSemesterDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "minimum": 1,
            "description": "해당 학기가 속한 연도 (예: 2024)",
            "example": 2024
          },
          "season": {
            "type": "string",
            "description": "해당 학기가 속한 계절 (Spring, Summer, Fall, Winter 중 하나)",
            "enum": [
              "Spring",
              "Summer",
              "Fall",
              "Winter"
            ],
            "example": "Summer"
          }
        }
      },
      "CreateLectureDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "Novice",
              "Advanced",
              "Expert"
            ],
            "description": "강의 난이도 (Novice, Intermediate, Advanced)"
          },
          "lectureNumber": {
            "type": "number",
            "minimum": 1,
            "description": "강의 회차수 (기본값: 10)"
          },
          "bojGroupId": {
            "type": "number",
            "description": "BOJ 그룹 ID"
          },
          "semesterId": {
            "type": "number",
            "minimum": 1,
            "description": "학기 ID (Semester의 id 값)"
          }
        },
        "required": [
          "level",
          "lectureNumber",
          "bojGroupId",
          "semesterId"
        ]
      },
      "LectureEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "고유 ID 값"
          },
          "level": {
            "type": "string",
            "enum": [
              "Novice",
              "Advanced",
              "Expert"
            ],
            "description": "강의 난이도. 초급, 중급, 고급 중 하나"
          },
          "lectureNumber": {
            "type": "number",
            "description": "강의의 총 회차수. 기본값은 10회차"
          },
          "bojGroupId": {
            "type": "number",
            "description": "강의에 연계된 BOJ 그룹 ID. 학생의 과제 출석체크를 위한 크롤링에 쓰임"
          },
          "semesterId": {
            "type": "number",
            "description": "강의가 속한 학기의 ID"
          }
        },
        "required": [
          "id",
          "level",
          "lectureNumber",
          "bojGroupId",
          "semesterId"
        ]
      },
      "UpdateLectureDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "Novice",
              "Advanced",
              "Expert"
            ],
            "description": "강의 난이도 (Novice, Intermediate, Advanced)"
          },
          "lectureNumber": {
            "type": "number",
            "minimum": 1,
            "description": "강의 회차수 (기본값: 10)"
          },
          "bojGroupId": {
            "type": "number",
            "description": "BOJ 그룹 ID"
          },
          "semesterId": {
            "type": "number",
            "minimum": 1,
            "description": "학기 ID (Semester의 id 값)"
          }
        }
      },
...
      "StudentEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "고유 ID 값"
          },
          "name": {
            "type": "string",
            "description": "학생의 이름"
          },
          "bojHandle": {
            "type": "string",
            "description": "학생의 BOJ 핸들 (문제풀이 체크 및 관리에 사용)"
          },
          "email": {
            "type": "string",
            "description": "학생의 이메일"
          },
          "phone": {
            "type": "string",
            "description": "학생의 전화번호"
          },
          "school": {
            "type": "string",
            "enum": [
              "YONSEI",
              "SOGANG",
              "HONGIK",
              "EWHA",
              "SOOKMYUNG"
            ],
            "description": "학생이 소속된 학교. 연세대, 서강대, 홍익대, 이화여대, 숙명여대 중 하나"
          },
          "studentNumber": {
            "type": "string",
            "description": "학생의 학번"
          }
        },
        "required": [
          "id",
          "name",
          "bojHandle",
          "email",
          "phone",
          "school",
          "studentNumber"
        ]
      },
...
      "CreateStudentLectureDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "학생 이름 (최대 50자)"
          },
          "bojHandle": {
            "type": "string",
            "description": "BOJ 핸들 (최대 50자, 고유값)"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "학생 이메일"
          },
          "phone": {
            "type": "string",
            "description": "학생 전화번호 (최대 20자)"
          },
          "school": {
            "type": "string",
            "enum": [
              "YONSEI",
              "SOGANG",
              "HONGIK",
              "EWHA",
              "SOOKMYUNG"
            ],
            "description": "학생 학교 (YONSEI, SOGANG, HONGIK, EWHA, SOOKMYUNG 중 하나)"
          },
          "studentNumber": {
            "type": "string",
            "description": "학생 학번 (최대 20자)"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "Refund",
              "NonRefund"
            ],
            "description": "학생 납부 상태 (PAID_30000, PAID_60000 중 하나)"
          },
          "refundAccount": {
            "type": "string",
            "description": "환불 계좌"
          },
          "lectureInfo": {
            "type": "object",
            "description": "강의 정보",
            "example": {
              "year": 2025,
              "season": "Winter",
              "level": "Novice"
            }
          },
          "studentLectureLogInfo": {
            "type": "object",
            "description": "학생 강의 수강 로그",
            "example": {
              "refundOption": "Refund",
              "refundAccount": "123-4567-890"
            }
          }
        },
        "required": [
          "name",
          "bojHandle",
          "email",
          "phone",
          "school",
          "studentNumber",
          "paymentStatus",
          "refundAccount",
          "lectureInfo",
          "studentLectureLogInfo"
        ]
      },
...
    }
  }
}